chore: Move spin_loader::oci::cache -> spin_loader::cache
