+++ crates/loader/src/cache.rs
