--- crates/loader/src/oci/mod.rs
@@ -1,5 +1,0 @@
-//! Functionality to get a prepared Spin application from an OCI registry.
-
-#![deny(missing_docs)]
-
-pub mod cache;
